#24 - moved Methods integration tests to MethodsIntegrationTest
