+++ src/test/java/nl/stil4m/mollie/ClientIntegrationTest.java
@@ -4,7 +4,6 @@ import com.google.common.collect.Sets;
 import nl.stil4m.mollie.domain.CreatePayment;
 import nl.stil4m.mollie.domain.CreatedPayment;
 import nl.stil4m.mollie.domain.Issuer;
-import nl.stil4m.mollie.domain.Method;
 import nl.stil4m.mollie.domain.Page;
 import nl.stil4m.mollie.domain.Refund;
 import org.junit.Before;
@@ -32,42 +31,6 @@ public class ClientIntegrationTest {
     public void before() throws InterruptedException {
         Thread.sleep(TEST_TIMEOUT);
         client = new ClientBuilder().withApiKey(VALID_API_KEY).build();
-    }
-
-    @Test
-    public void testGetMethods() throws IOException, URISyntaxException, InterruptedException {
-        ResponseOrError<Page<Method>> allResponse = client.methods().all(Optional.empty(), Optional.empty());
-        assertThat(allResponse.getSuccess(), is(true));
-
-        Page<Method> all = allResponse.getData();
-        assertThat(all.getCount(), is(10));
-        assertThat(all.getOffset(), is(0));
-        assertThat(all.getLinks(), is(notNullValue()));
-
-        Set<String> identifiers = all.getData().stream().map(Method::getId).collect(Collectors.toSet());
-        assertThat(identifiers.containsAll(Sets.newHashSet("ideal", "creditcard", "paypal")), is(true));
-    }
-
-    @Test
-    public void testGetMethod() throws IOException, URISyntaxException, InterruptedException {
-        ResponseOrError<Method> methodResponse = client.methods().get("ideal");
-        assertThat(methodResponse.getSuccess(), is(true));
-
-        Method method = methodResponse.getData();
-        assertThat(method.getId(), is("ideal"));
-        assertThat(method.getDescription(), is("iDEAL"));
-        assertThat(method.getAmount().getMinimum(), is(0.36));
-        assertThat(method.getAmount().getMaximum(), is(50000.0));
-        assertThat(method.getImage().getNormal(), is("https://www.mollie.com/images/payscreen/methods/ideal.png"));
-        assertThat(method.getImage().getBigger(), is("https://www.mollie.com/images/payscreen/methods/ideal@2x.png"));
-    }
-
-    @Test
-    public void testGetInvalidMethod() throws IOException, InterruptedException {
-        ResponseOrError<Method> methodResponse = client.methods().get("no-such-method");
-        assertThat(methodResponse.getSuccess(), is(false));
-        assertThat(methodResponse.getStatus(), is(404));
-        assertThat(methodResponse.getError().keySet(), is(Sets.newHashSet("error")));
     }
 
     @Test
